Rudimentary HUD code
The SpriteQueue thing from the Layer class has been broken into its own
tiny class so that one can be inserted in the Level class too, for HUD
purposes. Five hearts are now drawn to the screen. Five is a constant
number with no connection to the player object or anything else, for
now.