Textured backgrounds stop crashing
It was an old SFML-related bug I used to know about but had forgotten,
where letting sf::RenderTexture go out of scope is incredibly bad on
certain architectures, mine included. Fortunately, textured backgrounds
are always all the same size (256x256), so I was able to put in a sad
workaround with a static RenderTexture and conditionally calling its
Create method.