Create one Bunny per level
Now running some custom code in Hook_LevelLoad for creating a Bunny
object, rather than having Bunny be an object like any other and
assigned to the Jazz Start event. This prevents multiple bunnies from
being created in levels with more than one Jazz Start, or zero in levels
with none (but e.g. one or more MP Starts), and also happens to ensure
that the player sprites are drawn in front of everything else.

The actual method of picking which * Start event to use is somewhat
incomplete, but that's not a priority while your character is still
being assigned at random every time the game starts.