Draw point lighting
The level is now rendered based on the lighting buffer, which is cleared
to the property AmbientLightingLevel (currently always 0.25f, aka full
intensity) and then has point lights drawn to it. Their
sizes/brightnesses are still subject to revision; they're (obviously) a
port from the JJ2+ SFML branch, which is still unfinished.