Blur and fade the light buffer
to be more consistent with JJ2 behavior. Blurring is performed in
parallel by the GPU instead of linearly, meaning the thing where light
sources will affect light quantity below and to the right of them isn't
recreated, but that's good because that was dumb. Less rounding is
applied to flicker lights now because the blurring takes care of some of
that.