SpriteManager::SpriteTextures fix
Storing plain textures instead of unique_ptrs was a bad idea, because
when the vector got resized the textures were potentially moved and that
invalidated all previous pointers to them, in particular the ones in
animframes. This meant that when either sprites or lighting used more
than one texture, the majority of the time they were drawn using
incorrect or junk data.