Divide menu into separate screens
Well, in theory anyway--phenotypically the menu looks just the same as
it did before. But there are now MenuString and MenuScreen classes,
imported from the pertinent JJ2+ branch, and classes that inherit from
MenuScreen do the work of figuring out what's drawn and how the keys are
processed for the BunnyMenu.