Ice bullets freeze enemies
If you shoot a bee or normal turtle with an ice bullet, it will stop moving and be drawn using ice colors for a period of time. While frozen, it cannot hurt you (but it can be attacked by rabbits running at high speeds).

Only including the simple frozen shader for now, like JJ2(+) uses for resized/rotated sprites, not the full one that samples from its surroundings...